chore: remove test.mp4, make separate library for stdin_h264
