+++ src/streams/mod.rs
@@ -1,2 +1,5 @@
 #[cfg(target_os = "linux")]
 pub use v4l_h264::*;
+
+pub mod stdin_h264;
+pub use stdin_h264::*;
